Parallel Execution
Achieved Parallel Execution at Class and Method level. Please refer classes ParallelExecution1 and ParallelExecution2. Creation and using of the Driver object was bit difficult . FINALLY OVER CAME THAT issue . Happy I was coding at 4am in the morning to solve this issue . For few it may be minor but for me its a good achievement..........Kudos Zakir